Moved SPI library into macro and invoked from atmega328p
