+++ avr-hal-generic/src/spi.rs
@@ -1,80 +1,185 @@
-// //! SPI Implementation
+//! SPI Implementation
 
-// /// SPI Error
-// #[derive(Debug, Clone, Copy)]
-// pub enum Error { }
+/// Error type emitted by Spi in the event of a critical failure.  Errors have
+/// no information attached.
+#[derive(Debug, Clone, Copy)]
+pub enum SpiError {}
 
-// /// Implement traits for a SPI interface
-// #[macro_export]
-// macro_rules! impl_spi {
-//     (
-//         $(#[$spi_attr:meta])*
-//         pub struct $Spi:ident {
-//             peripheral: $SPI:ty,
-//             pins: {
-//                 // Might not need references to these pins?  Seems like r/w and clock are handled by hardware.
-//                 clock: $CLOCK:ident,
-//                 piso: $PISO:ident,
-//                 posi: $POSI:ident,
-//             },
-//             registers: {
-//                 control: $control:ident,
-//                 status: $status:ident,
-//                 data: $data:ident,
-//             },
-//         }
-//     ) => {
-//         $(#[$spi_attr])*
-//         pub struct $Spi {
-//             ss: $crate::hal::digital::v2::OutputPin,
-//             // TODO add necessary properties
-//         }
+/// Oscillator Clock Frequency division options.  Controls both SPR and SPI2X register bits.
+pub enum SerialClockRate {
+    OscfOver2,
+    OscfOver4,
+    OscfOver8,
+    OscfOver16,
+    OscfOver32,
+    OscfOver64,
+    OscfOver128,
+}
 
-//         impl $Spi
-//         {
-//             // TODO add settings arguments besides secondary select (optional?)
-//             /// Initialize the SPI peripheral
-//             pub fn new(ss: $crate::hal::digital::v2::OutputPin) -> $Spi {
-//                 // pull SS high
-//                 // store secondary-select, control, status, and register pins to struct
-//                 $Spi {}
-//             }
-//         }
+/// Order of data transmission, either MSB first or LSB first
+pub enum DataOrder {
+    MostSignificantFirst,
+    LeastSignificantFirst,
+}
 
-//         impl $crate::hal::spi::FullDuplex<u8> for $Spi {
-//             type Error = $crate::spi::Error;
+/// Polarity of clock (whether SCLK idles at low state or high state)
+pub enum SerialClockPolarity {
+    IdleHigh,
+    IdleLow,
+}
 
-//             fn write(&mut self, byte: u8) -> $crate::nb::Result<(), Self::Error> {
-//                 // I think it would be best to set all control bits for every write.  This way the user can have
-//                 // multiple Spi instances that communicate with different secondaries with no problem, even if they
-//                 // each have different settings.
-//                 // make sure the entire control register is set in one instruction for efficiency
-//                 // registers have modify/read/write/reset methods
+/// Clock sampling phase (check at leading or trailing edge of signal)
+pub enum SerialClockPhase {
+    SampleLeading,
+    SampleTrailing,
+}
 
-//                 // open communication with secondary via secondary-select pin
-//                 self.ss.set_low();
+/// Implement traits for a SPI interface
+#[macro_export]
+macro_rules! impl_spi {
+    (
+        $(#[$spi_attr:meta])*
+        pub struct $Spi:ident {
+            peripheral: $SPI:ty,
+            pins: {
+                posi: $POSI:ident,
+                piso: $PISO:ident,
+            }
+        }
+    ) => {
+        type POSI = $POSI<mode::Output>;
+        type PISO = $PISO<mode::Input<mode::PullUp>>;
 
-//                 // pull SS (instance of embedded_hal::serial::v2::OutputPin) low
-//                 // set SPIE (SPI enable) control bit to 1
-//                 // set MSTR (primary/secondary select) control bit to 1
+        /// Settings to pass to Spi.
+        ///
+        /// Easiest way to initialize is with
+        /// `Settings::default()`.  Otherwise can be instantiated with alternate
+        /// settings directly.
+        pub struct Settings {
+            data_order: DataOrder,
+            clock: SerialClockRate,
+            clock_polarity: SerialClockPolarity,
+            clock_phase: SerialClockPhase,
+        }
 
-//                 // set DORD (data order) control bit to user-defined setting (default 1)
-//                 // set CPOL (clock polarity) control bit to user-defined setting (default 0)
-//                 // set CPHA (clock phase) control bit to user-defined setting (default 0)
-//                 // set SPR (clock speed) control bits to user-defined setting (default 3)
-//                 // set SPIX2 (x2 clock speed) status bit to user-defined setting (default 0)
+        impl Default for Settings {
+            fn default() -> Self {
+                Settings {
+                    data_order: DataOrder::MostSignificantFirst,
+                    clock: SerialClockRate::OscfOver4,
+                    clock_polarity: SerialClockPolarity::IdleLow,
+                    clock_phase: SerialClockPhase::SampleTrailing,
+                }
+            }
+        }
 
-//                 // set $data to byte
+        /// Behavior for a SPI interface.
+        ///
+        /// Stores the SPI peripheral for register access.  In addition, it takes
+        /// ownership of the POSI and PISO pins to ensure they are in the correct mode.
+        /// Instantiate with the `new` method.
+        $(#[$spi_attr])*
+        pub struct $Spi {
+            peripheral: $SPI,
+            posi: POSI,
+            piso: PISO,
+            settings: Settings,
+        }
 
-//                 // close communication with secondary via secondary-select pin
-//                 self.ss.set_high();
-//                 Ok(())
-//             }
+        /// Implementation-specific behavior of the struct, including setup/tear-down
+        impl $Spi {
+            /// Instantiate an SPI with the registers, POSI/PISO pins, and settings
+            pub fn new(peripheral: $SPI, posi: POSI, piso: PISO, settings: Settings) -> $Spi {
+                Spi {
+                    peripheral,
+                    posi,
+                    piso,
+                    settings,
+                }
+            }
 
-//             fn read(&mut self) -> $crate::nb::Result<u8, Self::Error> {
-//                 // return and dereference $data
-//                 Ok(0)
-//             }
-//         }
-//     };
-// }
+            /// Release ownership of the peripheral and pins.  Instance can no-longer
+            /// be used after this is invoked.
+            pub fn release(self) -> ($SPI, POSI, PISO) {
+                (self.peripheral, self.posi, self.piso)
+            }
+
+            /// Write a byte to the data register, which begins transmission
+            /// automatically
+            fn write(&self, byte: u8) {
+                self.peripheral.spdr.write(|w| w.bits(byte));
+            }
+
+            /// Loop forever, checking the transmission complete bit until it is set
+            fn block_until_transfer_complete(&self) {
+                while self.peripheral.spsr.read().spif().bit_is_clear() { }
+            }
+
+            /// Sets up the control/status registers with the right settings for this secondary device
+            fn setup(&self) {
+                // set up control register
+                self.peripheral.spcr.write(|w| {
+                    // enable SPI
+                    w.spe().set_bit();
+                    // Set to primary mode
+                    w.mstr().set_bit();
+                    // set up data order control bit
+                    match self.settings.data_order {
+                        DataOrder::MostSignificantFirst => w.dord().clear_bit(),
+                        DataOrder::LeastSignificantFirst => w.dord().set_bit(),
+                    };
+                    // set up polarity control bit
+                    match self.settings.clock_polarity {
+                        SerialClockPolarity::IdleHigh => w.cpol().set_bit(),
+                        SerialClockPolarity::IdleLow => w.cpol().clear_bit(),
+                    };
+                    // set up phase control bit
+                    match self.settings.clock_phase {
+                        SerialClockPhase::SampleLeading => w.cpha().clear_bit(),
+                        SerialClockPhase::SampleTrailing => w.cpha().set_bit(),
+                    };
+                    // set up clock rate control bit
+                    match self.settings.clock {
+                        SerialClockRate::OscfOver2 => w.spr().val_0x00(),
+                        SerialClockRate::OscfOver4 => w.spr().val_0x00(),
+                        SerialClockRate::OscfOver8 => w.spr().val_0x01(),
+                        SerialClockRate::OscfOver16 => w.spr().val_0x01(),
+                        SerialClockRate::OscfOver32 => w.spr().val_0x02(),
+                        SerialClockRate::OscfOver64 => w.spr().val_0x02(),
+                        SerialClockRate::OscfOver128 => w.spr().val_0x03(),
+                    }
+                });
+                // set up 2x clock rate status bit
+                self.peripheral.spsr.write(|w| match self.settings.clock {
+                    SerialClockRate::OscfOver2 => w.spi2x().set_bit(),
+                    SerialClockRate::OscfOver4 => w.spi2x().clear_bit(),
+                    SerialClockRate::OscfOver8 => w.spi2x().set_bit(),
+                    SerialClockRate::OscfOver16 => w.spi2x().clear_bit(),
+                    SerialClockRate::OscfOver32 => w.spi2x().set_bit(),
+                    SerialClockRate::OscfOver64 => w.spi2x().clear_bit(),
+                    SerialClockRate::OscfOver128 => w.spi2x().set_bit(),
+                });
+            }
+        }
+
+        /// FullDuplex trait implementation, allowing this struct to be provided to
+        /// drivers that require it for operation.  Only 8-bit word size is supported
+        /// for now.
+        impl $crate::hal::spi::FullDuplex<u8> for $Spi {
+            type Error = $crate::spi::SpiError;
+
+            /// Sets up the device for transmission and sends the data
+            fn send(&mut self, byte: u8) -> $crate::nb::Result<(), Self::Error> {
+                self.setup();
+                self.write(byte);
+                self.block_until_transfer_complete();
+                Ok(())
+            }
+
+            /// Reads and returns the response in the data register
+            fn read(&mut self) -> $crate::nb::Result<u8, Self::Error> {
+                Ok(self.peripheral.spdr.read().bits())
+            }
+        }
+    };
+}
